Backend/file_writer_service/requirements.txt
Django==4.2.16
confluent-kafka==2.5.3
boto3==1.35.35
pyodbc>=4.0.39
pandas==2.3.0
influxdb-client==1.46.0
requests==2.32.3 
gunicorn>=23.0.0 
django-prometheus
psycopg2-binary==2.9.9
